requirements.txt
anyio==3.6.1
attrs==22.1.0
black==22.10.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.1
fastapi==0.79.1
firebase==3.0.1
firebase-admin==6.0.0
gcloud==0.18.3
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
h11==0.14.0
httplib2==0.20.4
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
jws==0.1.3
msgpack==1.0.4
mypy-extensions==0.4.3
oauth2client==4.1.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
proto-plus==1.22.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
pydantic==1.9.2
pyparsing==3.0.9
Pyrebase4==4.5.0
pytest==7.2.0
pytest-cov==4.0.0
python-dotenv==0.21.0
python-jwt==2.0.1
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
